+++ CHANGELOG.md
@@ -140,6 +140,7 @@
 - Accounting/Price tool: added CryptoCompare asset IDs to allow custom mapping of asset symbols.
 - BlockFi parser: added "BIA Deposit" transaction type.
 - Config: added USDC to crypto_list.
+- CoinTracker parser: added new export format. ([#401](https://github.com/BittyTax/BittyTax/issues/401))
 - Conversion tool: added (--cryptoasset) option.
 ### Changed
 - Conversion tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))